# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CoronaMecca API",
        "description": "CoronaMecca API.",
        "contact": {
            "email": "dann4624@edu.sde.dk"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://vacapi.semeicardia.online/api"
        }
    ],
    "paths": {
...
        "/boxLogs": {
            "get": {
                "tags": [
                    "Box Logs"
                ],
                "summary": "Get a list of Box Logs",
                "description": "Get a list of Box Logs",
                "operationId": "BoxLogsList",
                "responses": {
                    "200": {
                        "description": "List of Box Logs"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Box Logs"
                ],
                "summary": "Create a Box Log",
                "description": "Create a Box Log",
                "operationId": "BoxLogsCreate",
                "parameters": [
                    {
                        "name": "log_action_id",
                        "in": "query",
                        "description": "ID of the Log Action",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    },
                    {
                        "name": "box_id",
                        "in": "query",
                        "description": "ID of the Box",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "ID of the User",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    },
                    {
                        "name": "zone_id",
                        "in": "query",
                        "description": "ID of the Zone",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    },
                    {
                        "name": "position_id",
                        "in": "query",
                        "description": "ID of the Position",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Box Log created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/boxLogs/deleted": {
            "get": {
                "tags": [
                    "Box Logs"
                ],
                "summary": "Get a list of deleted box logs",
                "description": "Get a list of deleted box logs",
                "operationId": "BoxLogsListDeleted",
                "responses": {
                    "200": {
                        "description": "List of deleted Box Logs"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/boxLogs/{id}": {
            "get": {
                "tags": [
                    "Box Logs"
                ],
                "summary": "Update a Box Log",
                "description": "Update a Box Log",
                "operationId": "BoxLogsShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Box Log",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Box Log object"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Box Logs"
                ],
                "summary": "Update a Box Log",
                "description": "Update a Box Log",
                "operationId": "BoxLogsUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Box Log",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    },
                    {
                        "name": "log_action_id",
                        "in": "query",
                        "description": "ID of the Log Action",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    },
                    {
                        "name": "box_id",
                        "in": "query",
                        "description": "ID of the Box",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "ID of the User",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    },
                    {
                        "name": "zone_id",
                        "in": "query",
                        "description": "ID of the Zone",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    },
                    {
                        "name": "position_id",
                        "in": "query",
                        "description": "ID of the Position",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Box Log updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Box Logs"
                ],
                "summary": "Delete a Box Log",
                "description": "Delete a Box Log",
                "operationId": "BoxLogsDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Box Log",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Box Log deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/boxLogs/{id}/force": {
            "delete": {
                "tags": [
                    "Box Logs"
                ],
                "summary": "Permanently delete a Box Log",
                "description": "Permanently delete a Box Log",
                "operationId": "BoxLogsDeleteForce",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Box Log",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Box Log permanently deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/boxLogs/{id}/restore": {
            "put": {
                "tags": [
                    "Box Logs"
                ],
                "summary": "Restore a deleted Box Log",
                "description": "Restore a deleted Box Log",
                "operationId": "BoxLogsRestore",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Box Log",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Box Log restored"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}